[rlwe]: centralized linear transform
